+++ src/Hyperwallet.js
@@ -13,9 +13,10 @@ export default class Hyperwallet {
      * @param {string} config.username - The API username
      * @param {string} config.password - The API password
      * @param {string} [config.programToken] - The program token that is used for some API calls
+     * @param {Object} [config.encryptionData] - The JSON object of encryption data
      * @param {string} [config.server=https://api.sandbox.hyperwallet.com] - The API server to connect to
      */
-    constructor({ username, password, programToken, server = "https://api.sandbox.hyperwallet.com" }) {
+    constructor({ username, password, programToken, encryptionData, server = "https://api.sandbox.hyperwallet.com" }) {
         if (!username || !password) {
             throw new Error("You need to specify your API username and password!");
         }
@@ -26,7 +27,7 @@ export default class Hyperwallet {
          * @type {ApiClient}
          * @protected
          */
-        this.client = new ApiClient(username, password, server);
+        this.client = new ApiClient(username, password, server, encryptionData);
 
         /**
          * The program token that is used for some API calls
@@ -713,6 +714,133 @@ export default class Hyperwallet {
             throw new Error("paperCheckToken is required");
         }
         this.client.doGet(`users/${encodeURIComponent(userToken)}/paper-checks/${encodeURIComponent(paperCheckToken)}/status-transitions`, options, Hyperwallet.handle204Response(callback));
+    }
+
+    //--------------------------------------
+    // Transfers
+    //--------------------------------------
+
+    /**
+     * Create a transfer
+     *
+     * @param {Object} data - The transfer data
+     * @param {api-callback} callback - The callback for this call
+     */
+    createTransfer(data, callback) {
+        if (!data.sourceToken) {
+            throw new Error("sourceToken is required");
+        }
+        if (!data.destinationToken) {
+            throw new Error("destinationToken is required");
+        }
+        if (!data.clientTransferId) {
+            throw new Error("clientTransferId is required");
+        }
+        this.client.doPost("transfers", data, {}, callback);
+    }
+
+    /**
+     * Get a transfer
+     *
+     * @param {string} transferToken - The transfer token
+     * @param {api-callback} callback - The callback for this call
+     *
+     * @throws Will throw an error if transferToken is not provided
+     */
+    getTransfer(transferToken, callback) {
+        if (!transferToken) {
+            throw new Error("transferToken is required");
+        }
+        this.client.doGet(`transfers/${encodeURIComponent(transferToken)}`, {}, callback);
+    }
+
+    /**
+     * List all transfers
+     *
+     * @param {Object} options - The query parameters to send
+     * @param {api-callback} callback - The callback for this call
+     */
+    listTransfers(options, callback) {
+        this.client.doGet("transfers", options, Hyperwallet.handle204Response(callback));
+    }
+
+    /**
+     * Create a transfer status transition
+     *
+     * @param {string} transferToken - The transfer token
+     * @param {Object} data - The transfer status transition data
+     * @param {api-callback} callback - The callback for this call
+     * @throws Will throw an error if transferToken is not provided
+     */
+    createTransferStatusTransition(transferToken, data, callback) {
+        if (!transferToken) {
+            throw new Error("transferToken is required");
+        }
+
+        this.client.doPost(`transfers/${encodeURIComponent(transferToken)}/status-transitions`, data, {}, callback);
+    }
+
+    //--------------------------------------
+    // PayPal Accounts
+    //--------------------------------------
+
+    /**
+     * Create a PayPal account
+     *
+     * @param {string} userToken - The user token
+     * @param {Object} data - The PayPal account data
+     * @param {api-callback} callback - The callback for this call
+     *
+     * @throws Will throw an error if userToken is not provided
+     */
+    createPayPalAccount(userToken, data, callback) {
+        if (!userToken) {
+            throw new Error("userToken is required");
+        }
+        if (!data.transferMethodCountry) {
+            throw new Error("transferMethodCountry is required");
+        }
+        if (!data.transferMethodCurrency) {
+            throw new Error("transferMethodCurrency is required");
+        }
+        if (!data.email) {
+            throw new Error("email is required");
+        }
+        this.client.doPost(`users/${encodeURIComponent(userToken)}/paypal-accounts`, data, {}, callback);
+    }
+
+    /**
+     * Get a PayPal account
+     *
+     * @param {string} userToken - The user token
+     * @param {string} payPalAccountToken - The PayPal account token
+     * @param {api-callback} callback - The callback for this call
+     *
+     * @throws Will throw an error if userToken or payPalAccountToken is not provided
+     */
+    getPayPalAccount(userToken, payPalAccountToken, callback) {
+        if (!userToken) {
+            throw new Error("userToken is required");
+        }
+        if (!payPalAccountToken) {
+            throw new Error("payPalAccountToken is required");
+        }
+        this.client.doGet(`users/${encodeURIComponent(userToken)}/paypal-accounts/${encodeURIComponent(payPalAccountToken)}`, {}, callback);
+    }
+
+    /**
+     * List all PayPal accounts
+     *
+     * @param {string} userToken - The user token
+     * @param {Object} options - The query parameters to send
+     * @param {api-callback} callback - The callback for this call
+     * @throws Will throw an error if userToken is not provided
+     */
+    listPayPalAccounts(userToken, options, callback) {
+        if (!userToken) {
+            throw new Error("userToken is required");
+        }
+        this.client.doGet(`users/${encodeURIComponent(userToken)}/paypal-accounts`, options, Hyperwallet.handle204Response(callback));
     }
 
     //--------------------------------------